\name{AirlineDistances}
\alias{AirlineDistances}
\docType{data}
\title{
MMST AIRLINE DISTANCES DATA
}
\description{
airline distances, 464, 467, 481, 482, 484
}
\usage{data(AirlineDistances)}
\format{
A distance matrix of the distance in kilometers between the following 18
cities: Beijing, Cape Town, Hong Kong, Honolulu, London, Melbourne,
Mexico, Montreal, Moscow, New Delhi, New York, Paris, Rio de Janeiro,
Rome, San Francisco, Singapore, Stockholm, and Tokyo.
}
\details{
}
\source{
  National Geographic Society (1995), \emph{National Geographic Atlas of the World, Rev 6th Edition}, National Geograpic
}
\references{
  A. Izenman (2008), \emph{Modern Multivariate Statistical Techniques}, Springer
}
\examples{
}
\keyword{datasets}
