\name{alontop}
\alias{alontop}
\docType{data}
\title{
MMST ALONTOP DATA
}
\description{
colon cancer, 19, 20, 443, 444, 446
}
\usage{data(alontop)}
\format{
  A data frame with 62 observations (tissue samples) on 92 numeric
  variables (a subset of a larger set of more than 6500 genes).
}
\details{
}
\source{
}
\references{
  A. Izenman (2008), \emph{Modern Multivariate Statistical Techniques},
  Springer

  Alon, U., Barkai, N., Notterman, D., Gish, K., Ybarra, S., Mack, D.,
  and Levine, A. (1999). Broad patterns of gene expression revealed by
  clustering analysis of tumor and normal colon tissues probed by
  oligonucleotide arrays, \emph{Proceedings of the National Academy of
    Sciences}, \bold{96}, 6745-6750.
}
\examples{
}
\keyword{datasets}
