\name{baseball}
\alias{baseball}
\docType{data}
\title{
MMST BASEBALL DATA
}
\description{
Major League Baseball salaries, 307, 308, 309, 368
}
\usage{data(baseball)}
\format{
  A data frame with 337 observations on the following 18 variables.
  \describe{
    \item{\code{salary}}{a numeric vector}
    \item{\code{BA}}{a numeric vector}
    \item{\code{OBP}}{a numeric vector}
    \item{\code{Runs}}{a numeric vector}
    \item{\code{Hits}}{a numeric vector}
    \item{\code{X2B}}{a numeric vector}
    \item{\code{X3B}}{a numeric vector}
    \item{\code{HR}}{a numeric vector}
    \item{\code{RBI}}{a numeric vector}
    \item{\code{BB}}{a numeric vector}
    \item{\code{SO}}{a numeric vector}
    \item{\code{SB}}{a numeric vector}
    \item{\code{E}}{a numeric vector}
    \item{\code{FAE}}{a numeric vector}
    \item{\code{FA}}{a numeric vector}
    \item{\code{AE}}{a numeric vector}
    \item{\code{A}}{a numeric vector}
    \item{\code{Name}}{a character vector}
  }
}
\details{
}
\source{
}
\references{
  A. Izenman (2008), \emph{Modern Multivariate Statistical Techniques},
  Springer

  Watnik, M.R. (1998). Pay for play: are baseball salaries based on
  performance? \emph{Journal of Statistics Education}, \bold{6}.
}
\examples{
}
\keyword{datasets}
