\name{bodyfat}
\alias{bodyfat}
\docType{data}
\title{
MMST BODYFAT DATA
}
\description{
bodyfat, 116, 125, 128, 146, 148, 150, 151, 154
}
\usage{data(bodyfat)}
\format{
  A data frame with 252 observations on the following 15 variables.
  \describe{
    \item{\code{density}}{a numeric vector}
    \item{\code{bodyfat}}{a numeric vector}
    \item{\code{age}}{a numeric vector}
    \item{\code{weight}}{a numeric vector}
    \item{\code{height}}{a numeric vector}
    \item{\code{neck}}{a numeric vector}
    \item{\code{chest}}{a numeric vector}
    \item{\code{abdomen}}{a numeric vector}
    \item{\code{hip}}{a numeric vector}
    \item{\code{thigh}}{a numeric vector}
    \item{\code{knee}}{a numeric vector}
    \item{\code{ankle}}{a numeric vector}
    \item{\code{biceps}}{a numeric vector}
    \item{\code{forearm}}{a numeric vector}
    \item{\code{wrist}}{a numeric vector}
  }
}
\details{
}
\source{
  \url{http://lib.stat.cmu.edu/datasets/}
}
\references{
  A. Izenman (2008), \emph{Modern Multivariate Statistical Techniques}, Springer
}
\examples{
}
\keyword{datasets}
