\name{boston}
\alias{boston}
\docType{data}
\title{
MMST BOSTON HOUSING DATA
}
\description{
Boston housing, 158
}
\usage{data(boston)}
\format{
  A data frame with 506 observations on the following 14 variables.
  \describe{
    \item{\code{crim}}{a numeric vector}
    \item{\code{zn}}{a numeric vector}
    \item{\code{indus}}{a numeric vector}
    \item{\code{chas}}{a numeric vector}
    \item{\code{nox}}{a numeric vector}
    \item{\code{rm}}{a numeric vector}
    \item{\code{age}}{a numeric vector}
    \item{\code{dis}}{a numeric vector}
    \item{\code{rad}}{a numeric vector}
    \item{\code{tax}}{a numeric vector}
    \item{\code{ptratio}}{a numeric vector}
    \item{\code{black}}{a numeric vector}
    \item{\code{lstat}}{a numeric vector}
    \item{\code{medv}}{a numeric vector}
  }
}
\details{
}
\source{
  \url{http://lib.stat.cmu.edu/datasets/}
}
\references{
  A. Izenman (2008), \emph{Modern Multivariate Statistical Techniques}, Springer
}
\examples{
}
\keyword{datasets}
