\name{BritishTowns}
\alias{BritishTowns}
\docType{data}
\title{
MMST BRITISHTOWNS DATA
}
\description{
British towns, 504
}
\usage{data(BritishTowns)}
\format{
A distance matrix of the distances in kilometers between the
following 48 British towns: Aberdeen, Aberystwyth,
Barnstaple,Birmingham, Brighton, Bristol, Cambridge, Cardiff, Carlisle,
Carmarthen, Colchester, Dorchester, Dov, Edinburgh, Exeter,
Fort.William, Glasgow, Gloucester, Guildford, He, Holyhead, Hull,
Inverness, Kendal, Leeds, Lincoln, Liverpool, Maidstone, Manchester,
Middlesborough, Newcastle, Northampton, Norwich, Nottingham, Oxford,
Penzance, Perth, Plymouth, Preston, Salisbury, Sheffield, Shre,
Southampton, Stoke, Stranraer, Taunton, York, and London.
}
\details{
}
\source{
}
\references{
  A. Izenman (2008), \emph{Modern Multivariate Statistical Techniques}, Springer
}
\examples{
}
\keyword{datasets}
