\name{bupa}
\alias{bupa}
\docType{data}
\title{
MMST BUPA LIVER DISORDERS DATA
}
\description{
BUPA liver disorders, 258, 260, 348, 387, 508
}
\usage{data(bupa)}
\format{
  A data frame with 345 observations on the following 7 variables.
  \describe{
    \item{\code{mcv}}{a numeric vector}
    \item{\code{alkphos}}{a numeric vector}
    \item{\code{sgpt}}{a numeric vector}
    \item{\code{sgot}}{a numeric vector}
    \item{\code{gammagt}}{a numeric vector}
    \item{\code{drinks}}{a numeric vector}
    \item{\code{group}}{a factor with levels \code{1} \code{2}}
  }
}
\details{
}
\source{
  \url{http://archive.ics.uci.edu/ml/datasets.html}
}
\references{
  A. Izenman (2008), \emph{Modern Multivariate Statistical Techniques}, Springer
}
\examples{
}
\keyword{datasets}
