\name{cleveland}
\alias{cleveland}
\docType{data}
\title{
MMST CLEVELAND DATA
}
\description{
Cleveland heart-disease, 284, 286, 287, 289, 291, 314, 368
}
\usage{data(cleveland)}
\format{
  A data frame with 296 observations on the following 15 variables.
  \describe{
    \item{\code{age}}{a numeric vector}
    \item{\code{gender}}{a factor with levels \code{fem} \code{male}}
    \item{\code{cp}}{a factor with levels \code{abnang} \code{angina} \code{asympt} \code{notang}}
    \item{\code{trestbps}}{a numeric vector}
    \item{\code{chol}}{a numeric vector}
    \item{\code{fbs}}{a factor with levels \code{fal} \code{true}}
    \item{\code{restecg}}{a factor with levels \code{abn} \code{hyp} \code{norm}}
    \item{\code{thalach}}{a numeric vector}
    \item{\code{exang}}{a factor with levels \code{fal} \code{true}}
    \item{\code{oldpeak}}{a numeric vector}
    \item{\code{slope}}{a factor with levels \code{down} \code{flat} \code{up}}
    \item{\code{ca}}{a numeric vector}
    \item{\code{thal}}{a factor with levels \code{fix} \code{norm} \code{rev}}
    \item{\code{diag1}}{a factor with levels \code{buff} \code{sick}}
    \item{\code{diag2}}{a factor with levels \code{H} \code{S1} \code{S2} \code{S3} \code{S4}}
  }
}
\details{
}
\source{
  \url{http://archive.ics.uci.edu/ml/datasets.html}
}
\references{
  A. Izenman (2008), \emph{Modern Multivariate Statistical Techniques}, Springer
}
\examples{
}
\keyword{datasets}
