\name{color.stimuli}
\alias{color.stimuli}
\docType{data}
\title{
MMST COLOR.STIMULI DATA
}
\description{
perceptions of color, 468, 469, 503
}
\usage{data(color.stimuli)}
\format{
A distance matrix of colors represented by the following wavelengths (\emph{micrometers}):
434, 445, 465, 472, 490, 504, 537, 555, 584, 600, 610, 628, 651, and
674.
}
\details{
}
\source{
}
\references{
  A. Izenman (2008), \emph{Modern Multivariate Statistical Techniques},
  Springer

  Ekman, G. (1954). Dimensions of color vision, \emph{Journal of Psychology},
  \bold{38}, 467-474.
  
}
\examples{
}
\keyword{datasets}
