\name{COMBO17}
\alias{COMBO17}
\docType{data}
\title{
  MMST COMBO17 DATA
}
\description{
COMBO-17 galaxy photometric catalogue, 216, 219, 235
}
\usage{data(COMBO17)}
\format{
  A data frame with 3462 observations on 65 numeric variables.
}
\details{
}
\source{
}
\references{
  A. Izenman (2008), \emph{Modern Multivariate Statistical Techniques},
  Springer

  Wolf. C., Meisenheimer, M., Kleinheinrich, M., Borch, A., Dye, S.,
  Gray, M., Wisotski, L., Bell, E.F., Rix, H.-W., Cimatti, A., Hasinger,
  G., and Szokoly, G. (2004). A catalogue of the Chandra Deep Field
  South with multi-colour classification and photometric redshifts from
  COMBO-17, \emph{Astronomy & Astrophysics}, \code{arXiv:astro-ph/0403666v1}.
}
\examples{
}
\keyword{datasets}
