\name{covertype}
\alias{covertype}
\docType{data}
\title{
MMST COVERTYPE DATA}
\description{
covertype, 279
}
\usage{data(covertype)}
\format{
  A data frame with 581012 observations on 55 variables.
}
\details{
}
\source{
  \url{http://kdd.ics.uci.edu/databases/covertype/}
}
\references{
  A. Izenman (2008), \emph{Modern Multivariate Statistical Techniques}, Springer
}
\examples{
}
\keyword{datasets}
