\name{detergent}
\alias{detergent}
\docType{data}
\title{
MMST DETERGENT DATA}
\description{
laundry detergent, 156
}
\usage{data(detergent)}
\format{
  A data frame with 12 observations on 1173 variables. There are five Y
  variables, representing four compounds in an aquaeous solution (the
  fifth Y variable is the amount of water in the solution). The X input
  variables consist of mid-infrared spectrum values recorded as the
  absorbances at 1168 equally spaced frequencies in the detergent.
}
\details{
}
\source{
}
\references{
  A. Izenman (2008), \emph{Modern Multivariate Statistical Techniques}, Springer
}
\examples{
}
\keyword{datasets}
