\name{diabetes}
\alias{diabetes}
\docType{data}
\title{
MMST DIABETES DATA
}
\description{
diabetes, 272, 278, 348, 391
}
\usage{data(diabetes)}
\format{
  A data frame with 145 observations on the following 6 variables.
  \describe{
    \item{\code{glucose.area}}{a numeric vector}
    \item{\code{insulin.area}}{a numeric vector}
    \item{\code{SSPG}}{a numeric vector}
    \item{\code{relative.weight}}{a numeric vector}
    \item{\code{fasting.plasma.glucose}}{a numeric vector}
    \item{\code{class}}{a numeric vector}
  }
}
\details{
}
\source{
  \url{http://lib.stat.cmu.edu/datasets/Andrews/}
}
\references{
  A. Izenman (2008), \emph{Modern Multivariate Statistical Techniques},
  Springer

  Andrews, D.F. and Herzberg, A.M. (1985). \emph{Data}, New York: Springer.
}
\examples{
}
\keyword{datasets}
