\name{ecoli}
\alias{ecoli}
\docType{data}
\title{
MMST ECOLI DATA
}
\description{
e-coli, 273, 279, 348, 391
}
\usage{data(ecoli)}
\format{
  A data frame with 336 observations on the following 9 variables.
  \describe{
    \item{\code{label}}{a character vector}
    \item{\code{mvg}}{a numeric vector}
    \item{\code{gvh}}{a numeric vector}
    \item{\code{lip}}{a numeric vector}
    \item{\code{chg}}{a numeric vector}
    \item{\code{aac}}{a numeric vector}
    \item{\code{alm1}}{a numeric vector}
    \item{\code{alm2}}{a numeric vector}
    \item{\code{site}}{a factor with levels \code{cp} \code{im} \code{imL} \code{imS} \code{imU} \code{om} \code{omL} \code{pp}}
  }
}
\details{
}
\source{
  \url{http://archive.ics.uci.edu/ml/datasets.html}
}
\references{
  A. Izenman (2008), \emph{Modern Multivariate Statistical Techniques}, Springer
}
\examples{
}
\keyword{datasets}
