\name{food}
\alias{food}
\docType{data}
\title{
MMST FOOD DATA
}
\description{
nutritional value of food, 196, 198, 206, 208, 462, 612, 613, 631
}
\usage{data(food)}
\format{
  A data frame with 961 observations on the following 7 variables.
  \describe{
    \item{\code{Fat.grams}}{a numeric vector}
    \item{\code{Food.energy.calories}}{a numeric vector}
    \item{\code{Carbohydrates.grams}}{a numeric vector}
    \item{\code{Protein.grams}}{a numeric vector}
    \item{\code{Cholesterol.mg}}{a numeric vector}
    \item{\code{weight.grams}}{a numeric vector}
    \item{\code{Saturated.fat.grams}}{a numeric vector}
  }
}
\details{
}
\source{
  \url{http://www.ntwrks.com/~mikev/chart1.html}
}
\references{
  A. Izenman (2008), \emph{Modern Multivariate Statistical Techniques}, Springer
}
\examples{
}
\keyword{datasets}
