\name{geyser}
\alias{geyser}
\docType{data}
\title{
MMST GEYSER DATA
}
\description{
Old Faithful Geyser, 99, 100, 409, 410
}
\usage{data(geyser)}
\format{
  A data frame with 107 observations on the following 2 variables.
  \describe{
    \item{\code{X1}}{a numeric vector}
    \item{\code{X2}}{a numeric vector}
  }
}
\details{
}
\source{
}
\references{
  A. Izenman (2008), \emph{Modern Multivariate Statistical Techniques},
  Springer

  Weisberg, S. (1985). \emph{Applied Linear Regression, Second Edition}, New York: Wiley.
}
\examples{
}
\keyword{datasets}
