\name{gilgaied.soil}
\alias{gilgaied.soil}
\docType{data}
\title{
MMST GILGAIED SOIL DATA
}
\description{
gilgaied soil, 271, 278, 367
}
\usage{data(gilgaied.soil)}
\format{
  A data frame with 48 observations on the following 11 variables.
  \describe{
    \item{\code{pH}}{a numeric vector}
    \item{\code{N}}{a numeric vector}
    \item{\code{BD}}{a numeric vector}
    \item{\code{P}}{a numeric vector}
    \item{\code{Ca}}{a numeric vector}
    \item{\code{Mg}}{a numeric vector}
    \item{\code{K}}{a numeric vector}
    \item{\code{Na}}{a numeric vector}
    \item{\code{cond}}{a numeric vector}
    \item{\code{Block.no.}}{a factor with levels \code{1} \code{2} \code{3} \code{4}}
    \item{\code{Group.no.}}{a factor with levels \code{1} \code{2} \code{3} \code{4} \code{5} \code{6} \code{7} \code{8} \code{9} \code{10} \code{11} \code{12}}
  }
}
\details{
}
\source{
}
\references{
  A. Izenman (2008), \emph{Modern Multivariate Statistical Techniques},
  Springer

  Horton, I.F., Russell, J.S., and Moore,
  A.W. (1968). Multivariatecovariance and canonical analysis: a method
  for selecting the most effective discriminators in a multivariate
  situation, \emph{Biometrics}, \bold{24}, 845-858.
}
\examples{
}
\keyword{datasets}
