\name{glass}
\alias{glass}
\docType{data}
\title{
MMST GLASS DATA
}
\description{
forensic glass, 273, 348, 391, 508, 550
}
\usage{data(glass)}
\format{
  A data frame with 214 observations on the following 10 variables.
  \describe{
    \item{\code{RI}}{a numeric vector}
    \item{\code{Na}}{a numeric vector}
    \item{\code{Mg}}{a numeric vector}
    \item{\code{Al}}{a numeric vector}
    \item{\code{Si}}{a numeric vector}
    \item{\code{K}}{a numeric vector}
    \item{\code{Ca}}{a numeric vector}
    \item{\code{Ba}}{a numeric vector}
    \item{\code{Fe}}{a numeric vector}
    \item{\code{type}}{a factor with levels \code{1} \code{2} \code{3} \code{5} \code{6} \code{7}}
  }
}
\details{
}
\source{
  \url{http://archive.ics.uci.edu/ml/datasets.html}
}
\references{
  A. Izenman (2008), \emph{Modern Multivariate Statistical Techniques}, Springer
}
\examples{
}
\keyword{datasets}
