\name{Hidalgo1872}
\alias{Hidalgo1872}
\docType{data}
\title{
MMST HIDALGO 1872 STAMP DATA
}
\description{
Hidalgo postage stamps, 93, 96, 98
}
\usage{data(Hidalgo1872)}
\format{
  A data frame with 485 observations on the following 3 variables.
  \describe{
    \item{\code{thickness}}{a numeric vector}
    \item{\code{thicknessA}}{a numeric vector}
    \item{\code{thicknessB}}{a numeric vector}
  }
}
\details{
}
\source{
}
\references{
  A. Izenman (2008), \emph{Modern Multivariate Statistical Techniques},
  Springer

  Izenman, A.J. and Sommer, C.J. (1988). Philatelic mixtures and
  multimodal densities, \emph{Journal of the American Statistical Association},
  \bold{83}, 941-953.
}
\examples{
}
\keyword{datasets}
