\name{ionosphere}
\alias{ionosphere}
\docType{data}
\title{
MMST IONOSPHERE DATA
}
\description{
ionosphere, 258, 260, 348, 387
}
\usage{data(ionosphere)}
\format{
A data frame with 351 observations on 34 continuous variables and 1 factor, classifying the observation as \code{Good} (show some type of structure in the ionosphere) or \code{Bad} (pass through the ionosphere).
}
\details{
}
\source{
  \url{http://archive.ics.uci.edu/ml/datasets.html}
}
\references{
  A. Izenman (2008), \emph{Modern Multivariate Statistical Techniques}, Springer
}
\examples{
}
\keyword{datasets}
