\name{iris}
\alias{iris}
\docType{data}
\title{
MMST IRIS DATA
}
\description{
iris, 235, 274, 278, 348, 391
}
\usage{data(iris)}
\format{
  A data frame with 150 observations on the following 5 variables. This is R.A. Fisher's classic data set.
  \describe{
    \item{\code{sepal.length}}{a numeric vector}
    \item{\code{sepal.width}}{a numeric vector}
    \item{\code{petal.length}}{a numeric vector}
    \item{\code{petal.width}}{a numeric vector}
    \item{\code{type}}{a factor with levels \code{Iris-setosa} \code{Iris-versicolor} \code{Iris-virginica}}
  }
}
\details{
}
\source{
  \url{http://archive.ics.uci.edu/ml/datasets.html}
}
\references{
  A. Izenman (2008), \emph{Modern Multivariate Statistical Techniques}, Springer
}
\examples{
}
\keyword{datasets}
