\name{leukemia}
\alias{leukemia}
\docType{data}
\title{
MMST LEUKEMIA DATA
}
\description{
leukemia (ALL/AML), 451, 453, 461
}
\usage{data(leukemia)}
\format{
  A data frame with 72 observations on 7140 variables.
}
\details{
}
\source{
}
\references{
  A. Izenman (2008), \emph{Modern Multivariate Statistical Techniques},
  Springer

  Golub, T.R., Slonim, D.K., Tamayo, P., Huard, M., Gaasenbeek, J.P,,
  Mesirov, J.P., Coller, H., Loh, M.L., Downing, J.R., Caligiuri, M.A.,
  Bloomfield, C.D., and Lander, E.S. (1999). Molecular classification of
  cancer: class discovery and class prediction by gene expression
  monitoring, \emph{Science}, \bold{286}, 531-537.
}
\examples{
}
\keyword{datasets}
