\name{lloyd}
\alias{samp05}
\alias{samp25}
\alias{samp05d}
\alias{samp25d}
\docType{data}
\title{
MMST LLOYD'S BANK DATA
}
\description{
employee careers at Lloyds Bank, 477, 478, 489
}
\usage{
data(samp05)
data(samp25)
data(samp05d)
data(samp25d)
}
\format{
  There are two data frames and two distance matrices.
}
\details{
  There are four data sets utilized in this analysis, the \code{data} 
  function must be repeatedly used to load each of the four.
}
\source{
}
\references{
  A. Izenman (2008), \emph{Modern Multivariate Statistical Techniques}, Springer
  
  Stovel, K., Savage, M., and Bearman, P. (1996). Ascription into
  Achievement: Models of career systems at Lloyds Bank, 1890-1970.
  \emph{American Journal of Sociology}, \bold{102}, 358-399.
}
\examples{
}
\keyword{datasets}
