\name{MEG}
\alias{MEG}
\docType{data}
\title{
MMST MEG DATA
}
\description{
identifying artifacts in MEG recordings, 569
}
\usage{data(MEG)}
\format{
  A data frame with 122 observations on 17730 unnamed variables.
}
\details{
}
\source{
}
\references{
  A. Izenman (2008), \emph{Modern Multivariate Statistical Techniques},
  Springer

  Vigario, R,, Jousmaki, V., Hamalainen, M., Hari, R., and Oja,
  E. (1998). Independent component analysis for identification of
  artifacts in magnetoencephalographic recordings, In: \emph{Advances in
  Neural Information Processing Systems}, \bold{10}, pp. 229-235, Cambridge, MA:
  MIT Press.
  
}
\examples{
}
\keyword{datasets}
