\name{MMST-package}
\alias{MMST-package}
\alias{MMST}
\docType{package}
\title{
  DATASETS FROM MMST
}
\description{
  Data sets for \emph{Modern Multivariate Statistical Techniques}, by
  A. Izenman (2008).
}
\details{
\tabular{ll}{
Package: \tab MMST\cr
Type: \tab Package\cr
Version: \tab 0.6-1\cr
Date: \tab 2010-07-04\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Keith Halbert <keith.a.halbert@temple.edu>
}
\references{
    A. Izenman (2008), \emph{Modern Multivariate Statistical Techniques}, Springer
}
\seealso{
  \code{\link{MMST.out}}
}
\examples{
}
\keyword{ datasets }
\keyword{ IO }
