\name{morse}
\alias{morse}
\docType{data}
\title{
MMST MORSE DATA
}
\description{
confusion of Morse-code signals, 469, 470, 503, 504
}
\usage{data(morse)}
\format{
A data frame with 36 numeric observations on variables representing each of the 36 alphanumeric characters.
}
\details{
}
\source{
}
\references{
  A. Izenman (2008), \emph{Modern Multivariate Statistical Techniques},
  Springer

  Rothkopf, E.Z. (1957). A measure of stimulus similarity and errors in
  some paired-associate learning, \emph{Journal of Experimental
    Psychology}, \bold{53}, 94-101.
}
\examples{
}
\keyword{datasets}
