\name{ncifinal}
\alias{ncifinal}
\docType{data}
\title{
MMST NCIFINAL DATA
}
\description{
National Cancer Institute, 461
}
\usage{data(ncifinal)}
\format{
  A data frame with 5244 observations on 62 variables.
}
\details{
}
\source{
}
\references{
  A. Izenman (2008), \emph{Modern Multivariate Statistical Techniques}, Springer
  
  \url{http://www.cancer.gov/}

}
\examples{
}
\keyword{datasets}
