\name{norwaypaper1}
\alias{norwaypaper1}
\docType{data}
\title{
MMST NORWAYPAPER1 DATA
}
\description{
Norwegian paper quality, 166, 167, 190, 193, 194
}
\usage{data(norwaypaper1)}
\format{
  A data frame with 29 observations on the following 22 variables.
  \describe{
    \item{\code{Y1}}{a numeric vector}
    \item{\code{Y2}}{a numeric vector}
    \item{\code{Y3}}{a numeric vector}
    \item{\code{Y4}}{a numeric vector}
    \item{\code{Y5}}{a numeric vector}
    \item{\code{Y6}}{a numeric vector}
    \item{\code{Y7}}{a numeric vector}
    \item{\code{Y8}}{a numeric vector}
    \item{\code{Y9}}{a numeric vector}
    \item{\code{Y10}}{a numeric vector}
    \item{\code{Y11}}{a numeric vector}
    \item{\code{Y12}}{a numeric vector}
    \item{\code{Y13}}{a numeric vector}
    \item{\code{X1}}{a numeric vector}
    \item{\code{X2}}{a numeric vector}
    \item{\code{X3}}{a numeric vector}
    \item{\code{X4}}{a numeric vector}
    \item{\code{X5}}{a numeric vector}
    \item{\code{X6}}{a numeric vector}
    \item{\code{X7}}{a numeric vector}
    \item{\code{X8}}{a numeric vector}
    \item{\code{X9}}{a numeric vector}
  }
}
\details{
}
\source{
  \url{http://lib.stat.cmu.edu/datasets/}
}
\references{
  A. Izenman (2008), \emph{Modern Multivariate Statistical Techniques},
  Springer

  Aldrin, M. (1996). Moderate projection pursuit regression for
  multivariate response data, \emph{Computational Statistics & Data Analysis},
  \bold{21}, 501-531.
  
}
\examples{
}
\keyword{datasets}
