\name{pendigits}
\alias{pendigits}
\docType{data}
\title{
MMST PENDIGITS DATA
}
\description{
pen-based handwritten digit recognition, 211, 234, 274, 348, 391, 631
}
\usage{data(pendigits)}
\format{
  A data frame with 10992 observations on 36 unnamed variables.
}
\details{
}
\source{
  \url{http://archive.ics.uci.edu/ml/datasets.html}
}
\references{
  A. Izenman (2008), \emph{Modern Multivariate Statistical Techniques}, Springer
}
\examples{
}
\keyword{datasets}
