\name{pet}
\alias{pet}
\docType{data}
\title{
MMST PET DATA
}
\description{
PET yarns, 130, 133, 134, 136, 137, 142, 144, 156
}
\usage{data(pet)}
\format{
  A data frame with 28 observations on 270 variables.
  }
\details{
}
\source{
}
\references{
  A. Izenman (2008), \emph{Modern Multivariate Statistical Techniques},
  Springer

  Swierenga, H., de Weijer, A.P., van Wijk, R.J., and Buydens,
  L.M.C. (1999). Stategy for constructing robust multivariate
  calibration models, \emph{Chemometrics and Intelligent Laboratory Systems},
  \bold{49}, 1-17.
}
\examples{
}
\keyword{datasets}
