\name{pima}
\alias{pima}
\docType{data}
\title{
MMST PIMA DATA
}
\description{
Pima Indians diabetes, 292, 294, 296, 298, 299, 301, 302, 314, 368, 549
}
\usage{data(pima)}
\format{
  A data frame with 532 observations on the following 9 variables.
  \describe{
    \item{\code{npregnant}}{a numeric vector}
    \item{\code{glucose}}{a numeric vector}
    \item{\code{diastolic.bp}}{a numeric vector}
    \item{\code{skinfold.thickness}}{a numeric vector}
    \item{\code{bmi}}{a numeric vector}
    \item{\code{pedigree}}{a numeric vector}
    \item{\code{age}}{a numeric vector}
    \item{\code{classdigit}}{a factor with levels \code{0} \code{1}}
    \item{\code{class}}{a factor with levels \code{diabetic} \code{normal}}
  }
}
\details{
}
\source{
  \url{http://archive.ics.uci.edu/ml/datasets.html}
}
\references{
  A. Izenman (2008), \emph{Modern Multivariate Statistical Techniques}, Springer
}
\examples{
}
\keyword{datasets}
