\name{primate.scapulae}
\alias{primate.scapulae}
\docType{data}
\title{
MMST PRIMATE.SCAPULAE DATA
}
\description{
primate scapulae, 274, 279, 280, 348, 391, 420, 421, 461
}
\usage{data(primate.scapulae)}
\format{
  A data frame with 105 observations on the following 11 variables.
  \describe{
    \item{\code{genus}}{a numeric vector}
    \item{\code{AD.BD}}{a numeric vector}
    \item{\code{AD.CD}}{a numeric vector}
    \item{\code{EA.CD}}{a numeric vector}
    \item{\code{Dx.CD}}{a numeric vector}
    \item{\code{SH.ACR}}{a numeric vector}
    \item{\code{EAD}}{a numeric vector}
    \item{\code{beta}}{a numeric vector}
    \item{\code{gamma}}{a numeric vector}
    \item{\code{class}}{a factor with levels \code{Gorilla} \code{Homo} \code{Hylobates} \code{Pan} \code{Pongo}}
    \item{\code{classdigit}}{a factor with levels \code{1} \code{2} \code{3} \code{4} \code{5}}
  }
}
\details{
}
\source{
}
\references{
  A. Izenman (2008), \emph{Modern Multivariate Statistical Techniques},
  Springer

  Ashton, K.H., Oxnard, C.E., and Spence, T.F. (1965). Scapular shape
  and primate classification, \emph{Proceeding of the Zoological Society,
  London}, \bold{145}, 125-142.
  
}
\examples{
}
\keyword{datasets}
