\name{psych24r}
\alias{psych24r}
\docType{data}
\title{
MMST PSYCH24R DATA
}
\description{
psychological tests, 587, 588, 595
}
\usage{data(psych24r)}
\format{
  A data frame with 301 observations on the following 31 variables.
  \describe{
    \item{\code{Case}}{a numeric vector}
    \item{\code{Sex}}{a factor with levels \code{F} \code{M}}
    \item{\code{Age}}{a numeric vector}
    \item{\code{Grp}}{a numeric vector}
    \item{\code{V1}}{a numeric vector}
    \item{\code{V2}}{a numeric vector}
    \item{\code{V3}}{a numeric vector}
    \item{\code{V4}}{a numeric vector}
    \item{\code{V5}}{a numeric vector}
    \item{\code{V6}}{a numeric vector}
    \item{\code{V7}}{a numeric vector}
    \item{\code{V8}}{a numeric vector}
    \item{\code{V9}}{a numeric vector}
    \item{\code{V10}}{a numeric vector}
    \item{\code{V11}}{a numeric vector}
    \item{\code{V12}}{a numeric vector}
    \item{\code{V13}}{a numeric vector}
    \item{\code{V14}}{a numeric vector}
    \item{\code{V15}}{a numeric vector}
    \item{\code{V16}}{a numeric vector}
    \item{\code{V17}}{a numeric vector}
    \item{\code{V18}}{a numeric vector}
    \item{\code{V19}}{a numeric vector}
    \item{\code{V20}}{a numeric vector}
    \item{\code{V21}}{a numeric vector}
    \item{\code{V22}}{a numeric vector}
    \item{\code{V23}}{a numeric vector}
    \item{\code{V24}}{a numeric vector}
    \item{\code{V25}}{a numeric vector}
    \item{\code{V26}}{a numeric vector}
    \item{\code{group}}{a factor with levels \code{GRANT} \code{PASTEUR}}
  }
}
\details{
}
\source{
  \url{http://www.psych.yorku.ca/friendly/lab/files/psy6140/data/psych24r.sas}
}
\references{
  A. Izenman (2008), \emph{Modern Multivariate Statistical Techniques}, Springer
}
\examples{
}
\keyword{datasets}
