\name{root.stocks}
\alias{root.stocks}
\docType{data}
\title{
MMST ROOT STOCKS DATA
}
\description{
root-stocks of apple trees, 193
}
\usage{data(root.stocks)}
\format{
  A data frame with 104 observations on the following 5 variables.
  \describe{
    \item{\code{type}}{a factor with levels \code{I} \code{II} \code{III} \code{IV} \code{IX} \code{V} \code{VI} \code{VII} \code{X} \code{XII} \code{XIII} \code{XV} \code{XVI}}
    \item{\code{Y1}}{a numeric vector}
    \item{\code{Y2}}{a numeric vector}
    \item{\code{Y3}}{a numeric vector}
    \item{\code{Y4}}{a numeric vector}
  }
}
\details{
}
\source{
  \url{http://lib.stat.cmu.edu/datasets/Andrews/}
}
\references{
  A. Izenman (2008), \emph{Modern Multivariate Statistical Techniques}, Springer
  
  Andrews, D.F. and Herzberg, A.M. (1985). \emph{Data}, New York: Springer.
}
\examples{
}
\keyword{datasets}
