\name{satimage}
\alias{satimage}
\docType{data}
\title{
MMST SATIMAGE DATA
}
\description{
Landsat satellite image, 428, 431, 436, 438, 461
}
\usage{data(satimage)}
\format{
  A data frame with 4435 observations on 37 variables.
}
\details{
}
\source{
  \url{http://www.liacc.up.pt/}
}
\references{
  A. Izenman (2008), \emph{Modern Multivariate Statistical Techniques}, Springer
}
\examples{
}
\keyword{datasets}
