\name{shoplifting}
\alias{shoplifting}
\docType{data}
\title{
MMST SHOPLIFTING DATA
}
\description{
shoplifting in the Netherlands, 634, 635, 646, 647
}
\usage{data(shoplifting)}
\format{
  A data frame with 18 observations on the following 13 variables.
  \describe{
    \item{\code{clothing}}{a numeric vector}
    \item{\code{accessories}}{a numeric vector}
    \item{\code{tobacco}}{a numeric vector}
    \item{\code{writing}}{a numeric vector}
    \item{\code{books}}{a numeric vector}
    \item{\code{records}}{a numeric vector}
    \item{\code{goods}}{a numeric vector}
    \item{\code{sweets}}{a numeric vector}
    \item{\code{toys}}{a numeric vector}
    \item{\code{jewelry}}{a numeric vector}
    \item{\code{perfume}}{a numeric vector}
    \item{\code{tools}}{a numeric vector}
    \item{\code{other}}{a numeric vector}
  }
}
\details{
}
\source{
}
\references{
  A. Izenman (2008), \emph{Modern Multivariate Statistical Techniques},
  Springer

  van der Heijden, P.G.M., de Falguerolles, A., and de Leeuw,
  J. (1989). A combined approach to contingency table analysis using
  correspondence analysis and log-linear analysis, \emph{Applied Statistics},
  \bold{38}, 249-292.
}
\examples{
}
\keyword{datasets}
