\name{shuttle}
\alias{shuttle}
\docType{data}
\title{
MMST SHUTTLE DATA
}
\description{
shuttle, 274, 348, 391
}
\usage{data(shuttle)}
\format{
  A data frame with 43500 observations on 10 unnamed numeric variables.
}
\details{
}
\source{
  \url{http://archive.ics.uci.edu/ml/datasets.html}
}
\references{
  A. Izenman (2008), \emph{Modern Multivariate Statistical Techniques}, Springer
}
\examples{
}
\keyword{datasets}
