\name{soldat}
\alias{soldat}
\docType{data}
\title{
MMST SOLDAT DATA
}
\description{
aqueous solubility in drug discovery, 514, 515
}
\usage{data(soldat)}
\format{
  A data frame with 5631 observations on 72 input variables and 1 output variable.
}
\details{
}
\source{
}
\references{
  A. Izenman (2008), \emph{Modern Multivariate Statistical Techniques},
  Springer

  Culp, M., Johnson, K., and Michailidis, G. (2006). \code{ada}: an \code{R} package
  for stochastic boosting, \emph{Journal of Statistical Software}, \bold{17}, 2.
}
\examples{
}
\keyword{datasets}
