\name{sonar}
\alias{sonar}
\docType{data}
\title{
MMST SONAR DATA
}
\description{
sonar, 259, 260, 348, 387
}
\usage{data(sonar)}
\format{
  A data frame with 208 observations on 61 unnamed variables.
}
\details{
}
\source{
  \url{http://archive.ics.uci.edu/ml/datasets.html}
}
\references{
  A. Izenman (2008), \emph{Modern Multivariate Statistical Techniques}, Springer
}
\examples{
}
\keyword{datasets}
