\name{spambase}
\alias{spambase}
\docType{data}
\title{
MMST SPAMBASE DATA
}
\description{
spambase, 259, 260, 278, 348, 385, 387, 512, 549
}
\usage{data(spambase)}
\format{
  A data frame with 4601 observations on the following 59 variables.
  \describe{
    \item{\code{make}}{a numeric vector}
    \item{\code{address}}{a numeric vector}
    \item{\code{all}}{a numeric vector}
    \item{\code{xd}}{a numeric vector}
    \item{\code{our}}{a numeric vector}
    \item{\code{over}}{a numeric vector}
    \item{\code{remove}}{a numeric vector}
    \item{\code{internet}}{a numeric vector}
    \item{\code{order}}{a numeric vector}
    \item{\code{mail}}{a numeric vector}
    \item{\code{receive}}{a numeric vector}
    \item{\code{will}}{a numeric vector}
    \item{\code{people}}{a numeric vector}
    \item{\code{report}}{a numeric vector}
    \item{\code{addresses}}{a numeric vector}
    \item{\code{free}}{a numeric vector}
    \item{\code{business}}{a numeric vector}
    \item{\code{email}}{a numeric vector}
    \item{\code{you}}{a numeric vector}
    \item{\code{credit}}{a numeric vector}
    \item{\code{your}}{a numeric vector}
    \item{\code{font}}{a numeric vector}
    \item{\code{x000}}{a numeric vector}
    \item{\code{money}}{a numeric vector}
    \item{\code{hp}}{a numeric vector}
    \item{\code{hpl}}{a numeric vector}
    \item{\code{george}}{a numeric vector}
    \item{\code{x650}}{a numeric vector}
    \item{\code{lab}}{a numeric vector}
    \item{\code{labs}}{a numeric vector}
    \item{\code{telnet}}{a numeric vector}
    \item{\code{x857}}{a numeric vector}
    \item{\code{data}}{a numeric vector}
    \item{\code{x415}}{a numeric vector}
    \item{\code{x85}}{a numeric vector}
    \item{\code{technology}}{a numeric vector}
    \item{\code{x1999}}{a numeric vector}
    \item{\code{parts}}{a numeric vector}
    \item{\code{pm}}{a numeric vector}
    \item{\code{direct}}{a numeric vector}
    \item{\code{cs}}{a numeric vector}
    \item{\code{meeting}}{a numeric vector}
    \item{\code{original}}{a numeric vector}
    \item{\code{project}}{a numeric vector}
    \item{\code{re}}{a numeric vector}
    \item{\code{edu}}{a numeric vector}
    \item{\code{table}}{a numeric vector}
    \item{\code{conference}}{a numeric vector}
    \item{\code{x.}}{a numeric vector}
    \item{\code{x..}}{a numeric vector}
    \item{\code{x...1}}{a numeric vector}
    \item{\code{x..1}}{a numeric vector}
    \item{\code{x..2}}{a numeric vector}
    \item{\code{x..3}}{a numeric vector}
    \item{\code{crla}}{a numeric vector}
    \item{\code{crll}}{a numeric vector}
    \item{\code{crrt}}{a numeric vector}
    \item{\code{classdigit}}{a factor with levels \code{0} \code{1}}
    \item{\code{class}}{a factor with levels \code{email} \code{spam}}
  }
}
\details{
}
\source{
  \url{http://archive.ics.uci.edu/ml/datasets.html}
}
\references{
  A. Izenman (2008), \emph{Modern Multivariate Statistical Techniques}, Springer
}
\examples{
}
\keyword{datasets}
