\name{steganography}
\alias{steganography}
\docType{data}
\title{
MMST STEGANOGRAPHY DATA
}
\description{
steganography, 344, 345
}
\usage{data(steganography)}
\format{
  A data frame with 1000 observations on 73 variables.
}
\details{
}
\source{
}
\references{
  A. Izenman (2008), \emph{Modern Multivariate Statistical Techniques},
  Springer

  Kahn, D. (1996). The history of steganography, \emph{Proceedings of
  Information Hiding, First International Workshop}, Cambridge, U.K.
}
\examples{
}
\keyword{datasets}
