\name{swiss.roll}
\alias{swiss.roll}
\docType{data}
\title{
MMST SWISS ROLL DATA
}
\description{
Swiss roll, 598, 617, 619, 620, 622, 623
}
\usage{data(swiss.roll)}
\format{
  A data frame with 20000 observations on the following 5 variables.
  \describe{
    \item{\code{X1}}{a numeric vector}
    \item{\code{X2}}{a numeric vector}
    \item{\code{X3}}{a numeric vector}
    \item{\code{Y1}}{a numeric vector}
    \item{\code{Y2}}{a numeric vector}
  }
}
\details{
}
\source{
}
\references{
  A. Izenman (2008), \emph{Modern Multivariate Statistical Techniques}, Springer
}
\examples{
}
\keyword{datasets}
