\name{SwissBankNotes}
\alias{SwissBankNotes}
\docType{data}
\title{
MMST SWISSBANKNOTES DATA
}
\description{
Swiss bank notes, 235
}
\usage{data(SwissBankNotes)}
\format{
  A data frame with 200 observations on the following 6 variables.
  \describe{
    \item{\code{length}}{a numeric vector}
    \item{\code{height.left}}{a numeric vector}
    \item{\code{height.right}}{a numeric vector}
    \item{\code{inner.lower}}{a numeric vector}
    \item{\code{inner.upper}}{a numeric vector}
    \item{\code{diagonal}}{a numeric vector}
  }
}
\details{
}
\source{
}
\references{
  A. Izenman (2008), \emph{Modern Multivariate Statistical Techniques}, Springer
}
\examples{
}
\keyword{datasets}
