\name{tobacco}
\alias{tobacco}
\docType{data}
\title{
  MMST TOBACCO DATA
}
\description{
  chemical composition of tobacco, 183, 187
}
\usage{data(tobacco)}
\format{
  A data frame with 25 observations on the following 9 variables.
  \describe{
    \item{\code{Y1.BurnRate}}{a numeric vector}
    \item{\code{Y2.PercentSugar}}{a numeric vector}
    \item{\code{Y3.PercentNicotine}}{a numeric vector}
    \item{\code{X1.PercentNitrogen}}{a numeric vector}
    \item{\code{X2.PercentChlorine}}{a numeric vector}
    \item{\code{X3.PercentPotassium}}{a numeric vector}
    \item{\code{X4.PercentPhosphorus}}{a numeric vector}
    \item{\code{X5.PercentCalcium}}{a numeric vector}
    \item{\code{X6.PercentMagnesium}}{a numeric vector}
  }
}
\details{
}
\source{
}
\references{
  A. Izenman (2008), \emph{Modern Multivariate Statistical Techniques},
  Springer

  Anderson, R.L. and Bancroft, T.A. (1952). \emph{Statistical Theory in
  Research}, New York: McGraw-Hill.
}
\examples{
}
\keyword{datasets}
