\name{tumors}
\alias{tumors}
\docType{data}
\title{
MMST TUMORS DATA
}
\description{
four childhood tumors, 541, 545, 550
}
\usage{data(tumors)}
\format{
  A data frame with 2308 observations on 90 variables.
}
\details{
}
\source{
  \url{http://research.nhgri.nih.gov/microarray/Supplement/}
}
\references{
  A. Izenman (2008), \emph{Modern Multivariate Statistical Techniques},
  Springer

  Khan, J., Wei, J.S., Ringner, M., Saal, L.H., Ladanyi, M., Westermann,
  F., Berthold, F., Schwab, M., Antonescu, C.R., Peterson, C., and
  Meltzer, P.S. (2001). Classification and diagnostic prediction of
  cancers using gene expression profiling and artificial neural
  networks, \emph{Nature Medicine}, \bold{7}, 673-679.
}
\examples{
}
\keyword{datasets}
