\name{turtles}
\alias{turtles}
\docType{data}
\title{
MMST TURTLES DATA
}
\description{
turtle carapaces, 234
}
\usage{data(turtles)}
\format{
  A data frame with 48 observations on the following 4 variables.
  \describe{
    \item{\code{sex}}{a factor with levels \code{f} \code{m}}
    \item{\code{length}}{a numeric vector}
    \item{\code{width}}{a numeric vector}
    \item{\code{height}}{a numeric vector}
  }
}
\details{
}
\source{
}
\references{
  A. Izenman (2008), \emph{Modern Multivariate Statistical Techniques}, Springer
}
\examples{
}
\keyword{datasets}
