\name{ushighways}
\alias{ushighways}
\docType{data}
\title{
MMST USHIGHWAYS DATA
}
\description{
U.S. highways, 106
}
\usage{data(ushighways)}
\format{
  A data frame with 221 observations on the following 4 variables.
  \describe{
    \item{\code{Interstate}}{a numeric vector}
    \item{\code{State}}{a factor with levels \code{AL} \code{AR} \code{CA} \code{CO} \code{CT} \code{DE} \code{FL} \code{GA} \code{IA} \code{ID} \code{IL} \code{IN} \code{KS} \code{KY} \code{LA} \code{MA} \code{MD} \code{ME} \code{MI} \code{MN} \code{MO} \code{MS} \code{NC} \code{NE} \code{NH} \code{NJ} \code{NY} \code{OH} \code{OK} \code{OR} \code{PA} \code{RI} \code{SC} \code{SD} \code{TN} \code{TX} \code{UT} \code{VA} \code{VT} \code{WA} \code{WI} \code{WV} \code{WY}}
    \item{\code{Approx.Miles}}{a numeric vector}
    \item{\code{Location}}{a character vector}
  }
}
\details{
}
\source{
}
\references{
  A. Izenman (2008), \emph{Modern Multivariate Statistical Techniques},
  Springer

  Rand McNally (1992), \emph{Rand McNally 1993 Business Traveler's Road
  Atlas and Guide to Major Cities}, Rand McNally
  
}
\examples{
}
\keyword{datasets}
