\name{vehicle}
\alias{vehicle}
\docType{data}
\title{
MMST VEHICLE DATA
}
\description{
vehicle, 274, 302, 304, 348, 391
}
\usage{data(vehicle)}
\format{
  A data frame with 564 observations on the following 20 variables.
  \describe{
    \item{\code{Comp}}{a numeric vector}
    \item{\code{Circ}}{a numeric vector}
    \item{\code{Dcirc}}{a numeric vector}
    \item{\code{RR}}{a numeric vector}
    \item{\code{PrAxisAR}}{a numeric vector}
    \item{\code{MaxLAR}}{a numeric vector}
    \item{\code{ScatterR}}{a numeric vector}
    \item{\code{Elong}}{a numeric vector}
    \item{\code{PrAxisRect}}{a numeric vector}
    \item{\code{MaxLRect}}{a numeric vector}
    \item{\code{SvarMajAxis}}{a numeric vector}
    \item{\code{SvarMinAxis}}{a numeric vector}
    \item{\code{SradGyration}}{a numeric vector}
    \item{\code{SkewMajAxis}}{a numeric vector}
    \item{\code{SkewMinAxis}}{a numeric vector}
    \item{\code{KurtMinAxis}}{a numeric vector}
    \item{\code{KurtMajAxis}}{a numeric vector}
    \item{\code{Hratio}}{a numeric vector}
    \item{\code{classdigit}}{a factor with levels \code{1} \code{2} \code{3} \code{4}}
    \item{\code{class}}{a factor with levels \code{bus} \code{opel} \code{saab} \code{van}}
  }
}
\details{
}
\source{
  \url{http://archive.ics.uci.edu/ml/datasets.html}
}
\references{
  A. Izenman (2008), \emph{Modern Multivariate Statistical Techniques}, Springer
}
\examples{
}
\keyword{datasets}
