\name{wdbc}
\alias{wdbc}
\docType{data}
\title{
MMST WDBC DATA
}
\description{
Wisconsin diagnostic breast cancer, 239, 241, 246, 247, 249, 251, 255, 256, 260, 279, 348, 387, 462, 508, 550
}
\usage{data(wdbc)}
\format{
  A data frame with 569 observations on the following 32 variables.
  \describe{
    \item{\code{id}}{a numeric vector}
    \item{\code{class}}{a factor with levels \code{B} \code{M}}
    \item{\code{radius.mv}}{a numeric vector}
    \item{\code{texture.mv}}{a numeric vector}
    \item{\code{peri.mv}}{a numeric vector}
    \item{\code{area.mv}}{a numeric vector}
    \item{\code{smooth.mv}}{a numeric vector}
    \item{\code{comp.mv}}{a numeric vector}
    \item{\code{scav.mv}}{a numeric vector}
    \item{\code{ncav.mv}}{a numeric vector}
    \item{\code{symt.mv}}{a numeric vector}
    \item{\code{fracd.mv}}{a numeric vector}
    \item{\code{radius.sd}}{a numeric vector}
    \item{\code{texture.sd}}{a numeric vector}
    \item{\code{peri.sd}}{a numeric vector}
    \item{\code{area.sd}}{a numeric vector}
    \item{\code{smooth.sd}}{a numeric vector}
    \item{\code{comp.sd}}{a numeric vector}
    \item{\code{scav.sd}}{a numeric vector}
    \item{\code{ncav.sd}}{a numeric vector}
    \item{\code{symt.sd}}{a numeric vector}
    \item{\code{fracd.sd}}{a numeric vector}
    \item{\code{radius.ev}}{a numeric vector}
    \item{\code{texture.ev}}{a numeric vector}
    \item{\code{peri.ev}}{a numeric vector}
    \item{\code{area.ev}}{a numeric vector}
    \item{\code{smooth.ev}}{a numeric vector}
    \item{\code{comp.ev}}{a numeric vector}
    \item{\code{scav.ev}}{a numeric vector}
    \item{\code{ncav.ev}}{a numeric vector}
    \item{\code{symt.ev}}{a numeric vector}
    \item{\code{fracd.ev}}{a numeric vector}
  }
}
\details{
}
\source{
}
\references{
  A. Izenman (2008), \emph{Modern Multivariate Statistical Techniques},
  Springer

  Street, W.N., Wolberg, W.H., and Mangasarian, O.L. (1993). Nuclear
  feature extraction for breast tumor diagnosis, \emph{IS&T/SPIE International
  Symposium on Electronic Imaging: Science and Technology} (San Jose,
  CA), \bold{1905}, 861-870.
}
\examples{
}
\keyword{datasets}
