\name{wine}
\alias{wine}
\docType{data}
\title{
MMST WINE DATA
}
\description{
wine, 275, 278, 348, 391
}
\usage{data(wine)}
\format{
  A data frame with 178 observations on the following 15 variables.
  \describe{
    \item{\code{Alcohol}}{a numeric vector}
    \item{\code{MalicAcid}}{a numeric vector}
    \item{\code{Ash}}{a numeric vector}
    \item{\code{AlcAsh}}{a numeric vector}
    \item{\code{Mg}}{a numeric vector}
    \item{\code{Phenols}}{a numeric vector}
    \item{\code{Flav}}{a numeric vector}
    \item{\code{NonFlavPhenols}}{a numeric vector}
    \item{\code{Proa}}{a numeric vector}
    \item{\code{Color}}{a numeric vector}
    \item{\code{Hue}}{a numeric vector}
    \item{\code{OD}}{a numeric vector}
    \item{\code{Proline}}{a numeric vector}
    \item{\code{classdigit}}{a factor with levels \code{1} \code{2} \code{3}}
    \item{\code{class}}{a factor with levels \code{Barbera} \code{Barolo} \code{Grignolino}}
  }
}
\details{
}
\source{
  \url{http://archive.ics.uci.edu/ml/datasets.html}
}
\references{
  A. Izenman (2008), \emph{Modern Multivariate Statistical Techniques}, Springer
}
\examples{
}
\keyword{datasets}
