\name{x498.matrix}
\alias{x498.matrix}
\docType{data}
\title{
MMST 498 MATRIX DATA
}
\description{
mapping the protein universe, 484, 486
}
\usage{data(x498.matrix)}
\format{
A distance matrix mapping 498 unnamed variables representing proteins.
}
\details{
}
\source{
}
\references{
  A. Izenman (2008), \emph{Modern Multivariate Statistical Techniques},
  Springer

  Hou, J., Sims, G.E., Zhang, C., and Kim, S.-H. (2003). A global
  representation of the protein fold space, \emph{Proceedings of the National
  Academy of Sciences}, \bold{100}, 2386-2390.
}
\examples{
}
\keyword{datasets}
