\name{yeast}
\alias{yeast}
\docType{data}
\title{
MMST YEAST DATA
}
\description{
yeast, 275, 279, 348, 391, 508
}
\usage{data(yeast)}
\format{
  A data frame with 1484 observations on the following 10 variables.
  \describe{
    \item{\code{yeast}}{a character vector}
    \item{\code{mcg}}{a numeric vector}
    \item{\code{gvh}}{a numeric vector}
    \item{\code{alm}}{a numeric vector}
    \item{\code{mit}}{a numeric vector}
    \item{\code{erl}}{a numeric vector}
    \item{\code{pox}}{a numeric vector}
    \item{\code{vac}}{a numeric vector}
    \item{\code{nuc}}{a numeric vector}
    \item{\code{site}}{a factor with levels \code{CYT} \code{ERL} \code{EXC} \code{ME1} \code{ME2} \code{ME3} \code{MIT} \code{NUC} \code{POX} \code{VAC}}
  }
}
\details{
}
\source{
  \url{http://archive.ics.uci.edu/ml/datasets.html}
}
\references{
  A. Izenman (2008), \emph{Modern Multivariate Statistical Techniques}, Springer
}
\examples{
}
\keyword{datasets}
